+++ backend/src/services/UserService.cc
@@ -13,7 +13,7 @@ std::optional<User> UserService::getUser(int64_t id) {
   return PgClient::findUserById(id);
 }
 
-std::vector<User> UserService::listUsers(int offset, int limit) {
+std::vector<User> UserService::listUsers(int64_t offset, int64_t limit) {
   return PgClient::listUsers(offset, limit);
 }
 
@@ -21,8 +21,7 @@ bool UserService::deleteUser(int64_t id) { return PgClient::deleteUser(id); }
 
 int64_t UserService::countAdmins() { return PgClient::countAdmins(); }
 
-std::optional<int64_t>
-UserService::countActiveLoansByUser(int64_t userId) {
+std::optional<int64_t> UserService::countActiveLoansByUser(int64_t userId) {
   return PgClient::countActiveLoanRecordsByUser(userId);
 }
